refactor: add feature entrypoints
