+++ src/app/(home)/work/[slug]/page.tsx
@@ -6,7 +6,7 @@ import { BlurImage } from '@/components/blur-image'
 import { mdxComponents } from '@/components/content/mdx/components'
 import { GitHubCode } from '@/components/content/mdx/github-code'
 import { TOCMinimap, TOCPopover } from '@/components/features/toc'
-import { PreviewButton } from '@/components/features/work/preview-button'
+import { PreviewButton } from '@/components/features/work'
 import { Icons } from '@/components/icons/icons'
 import { WorkJsonLd } from '@/components/json-ld'
 import { Section } from '@/components/layout/sections'
