+++ src/components/features/blog/index.ts
@@ -1,0 +1,2 @@
+export * from './post-card'
+export * from './tag-card'
