+++ src/components/features/contributions/index.ts
@@ -1,0 +1,1 @@
+export * from './contribution-graph'
