+++ src/components/features/work/index.ts
@@ -1,0 +1,3 @@
+export * from './preview-button'
+export * from './work-card'
+export * from './work-experience'
